Added missing features in python 3.9
