+++ README.md
@@ -1,17 +1,17 @@
 # Pangram Programs
 This is a collection of hololexical, symbol-exhaustive, and/or language-feature saturated codes implemented in various languages. This is where programming languages, semantic completionism, and maybe even a hint of poetry converge.
 
-Source code contributed to this repository should be what ["The quick brown fox jumped over the lazy dog."](https://en.wikipedia.org/wiki/The_quick_brown_fox_jumps_over_the_lazy_dog) is to English. For the uninitiated, this is a "pangram" and uses all 26 alphabetical English letter at least once.  But, what of English's operators? It features quite a lot of puncuation marks! Here is a punctional equivalent pangram for English:
->While he was writing his book, Magic: Avery & the Little Known World! he stated, “I always… wanted to ask a magishin [sic], ‘How do you do it? And will you teach me for $20?’”; he luckily got the chance while he was living above a shop (where he lived from 1993 — 1997, roughly 10% of his life).
+Source code contributed to this repository should be what ["The quick brown fox jumped over the lazy dog."](https://en.wikipedia.org/wiki/The_quick_brown_fox_jumps_over_the_lazy_dog) is to English. For the uninitiated, this is a "pangram" and uses all 26 alphabetical English letter at least once.  But, what of English's operators? It features quite a lot of punctuation marks! Here is a punctual equivalent pangram for English:
+>While he was writing his book, Magic: Avery & the Little Known World! he stated, “I always… wanted to ask a magician [sic], ‘How do you do it? And will you teach me for $20?’”; he luckily got the chance while he was living above a shop (where he lived from 1993 — 1997, roughly 10% of his life).
 
 <sub>_Slightly modified [this original version.](https://www.quora.com/What-is-the-English-Sentence-which-contains-all-14-punctuation-marks)_</sub>
 
 Now imagine trying to use every letter and punctuation mark in as short of prose as possible&mdash;quite a poetic challenge. And similarly, what's the shortest grammatically valid prose that uses all English words? Such a problem is a bit nebulous, but luckily we have languages which are unambiguous where such a solution is likely to exist: those which are computed.
 
-In essence, code here should be thought of as a tour of its respective language but with the fun of a challenge. Specifically, this may manifest as a script that uses all reserved words and functional operators in the language, every parsible symbol, every block structure, etc.
+In essence, code here should be thought of as a tour of its respective language but with the fun of a challenge. Specifically, this may manifest as a script that uses all reserved words and functional operators in the language, every parsable symbol, every block structure, etc.
 
 For the practically inclined, think of codes here as a demonstrative cheatsheet. Rather than a list of all functions and keywords, why not show their usage comprehensively in one piece of code? It may be as challenging as you please.
 
 The only requirement is that some aspect of the language the code should be executed with is comprehensively represented.
 
-For an example, please see [my python sample.](python/all-keywords-builtins-operators/README.md)
+For an example, please see [my python sample.](python/all-keywords-builtins-operators-3.9/README.md)
